Refactor partnership access to separate functions
In the cases of correct access (user belongs to the partner whose page
they're trying to view) and incorrect access (user does not belong to the
the partner), the flow has been moved to separate functions to allow
better readability.

To refactor these functions, the type StoryBlokData was created. This will
be pulled out at a later time.