Fix issue with password-reset flow
This commit fixes bug which redirects users to the login page if they
click on a reset password link in their email.

This bug occured because the redirect link from firebase was not identified
by the app as pointing to a public page. Instead, an auth-guarded component
was served which then triggered the app to show the login page.

Going deeper in the code, the issue occured because the mechanism to
identify public paths did not work when the path contained query params.
Since the reset-password email links did contain query params, it
was incorrectly classes as needing an auth guard.

The variables which hold the public paths and the current path have
been renamed to contain 'head' as this makes it clear to the reader
which part of the path is referred to in the code.

050422-1958 / Fix password reset flow