Wrap column field in array if necessary
After upgrading Storyblok from V1 to V2 the column prop for
StoryblokRow has changed.
Previously, the StoryblokRow column prop was always an array.
However now the prop can be an array or an object.

To handle this lack of consistency in the prop type, this commit
wraps the column prop in an array if it is an object.

070922-1750 / Fix build fail