Remove chat and course options
These two features will now be provided by automatic access codes.
The only time admin generated codes are required is for therapy.
The use of the `selectedTier` variable is still left in case we want to bring back multiple options in the future.

Add validation to redirect user when is logged in