Add new navigation drawer links component
This is required as the original navigation menu can no longer be used for the drawer. This is due to the new secondary navigation structure.

In a regular view, the new navigation menu only has a few items - courses, notes etc.. are in the second nav bar. In the mobile view, the nav drawer needs to now contain a lot more items including courses, notes, therapy - hence the creation of the new NavigationDrawerLinks component.