Add check on firebase user update to auth guard
This is needed because on a fresh load of a page (e.g. typing in <bloom-url>/admin/dashboard in address bar and pressing enter), the firebase user is empty and the state user token is null.
The auth guard treats this lack of a token as indicating the user has logged out and redirects to the login page.
However, this is incorrect as the user has not actually logged out - it only appears so briefly. Eventually on a fresh load, client side firebase will re-authorise and receive a new ID token.
Then in the _app file, there is a listener which listens for ID token change and updates the user state.
This will now set the user token to a value.

The above redirect issue happens because the auth guard token check ran before the firebase user and the user state were correctly updated.
This change will make the auth guard useEffect to check whether the firebase user has been updated.
Once the firebase user is updated, then the logic which checks for the token will be run.