Setting rarely used props as optional with default values
This will allow IBA specific course page which will be added in subsequent commits to not need to supply these props. The concept of live courses is not being used in Bloom currently though it maybe be brought back in the future. IBA will not be run as a live course so this change allows IBA specific pages to be simplified.