fix: pwa banner flicker
This commit fixes an issue where the usePWA hook
led to an initial falsy 'generic' bannerState.

The generic bannerState will actualy render the banner.
The 'hidden' bannerState which will prevent the rendering
was detected with a short delay - which led to shortly render
the banner (flicker).

The fix is to adjust the defaultBanner state to hidden instead
of generic + to explicit check for 'generic' bannerState conditions
to be true.